Title: 선두를 지켜라, Time: 262ms, Memory: 28MB, Status: Passed - CodeTree

+++ 241029/선두를 지켜라/keep-the-lead.java
@@ -38,8 +38,8 @@ public class Main {
             }
         }
 
-        int status = 0;
-        int tempStatus;
+        long status = 0;
+        long tempStatus;
         int cnt = 0;
         for (int i = 1; i <= timeA; i++) {
             tempStatus = timelineA[i] - timelineB[i];
